+++ packages/wikibase-query/RELEASENOTES.md
@@ -2,4 +2,7 @@
 
 ## Unreleased
 
-- Initial commit with a `WikibaseQueryService` and `SimpleQueryService` extracted from `wikibase-commands` package and refactored.
+- Initial version with:
+  - `WikibaseQueryService` extracted from `wikibase-commands` package and refactored.
+  - `SimpleQueryService` extracted from `wikibase-commands` package and refactored.
+  - `PrefixSets`